Active Directory New User Creation
This file will take a word document from the folder New Users on the USER Desktop and will grab the information from the table to fill in the blanks. This script is specific to the company I worked for. Additionally contained within files in the same folder is an excel document with more information that can be referenced and populated into the AD membership. This script was made to automate the process of creating New Users in Active Directory for my work. This saved the IT team at my work over 4 hours of time each week.